test(doctor): split state-integrity suite to satisfy max-lines
